Split execution_environment from package_dependencies, use a single list of PackageInfo
