+++ tests/core/usage_statistics/test_package_dependencies.py
@@ -1,240 +1,8 @@
 from typing import List
-from unittest import mock
 
 import pytest
-from packaging import version
 
-from great_expectations.core.usage_statistics.package_dependencies import (
-    GEDependencies,
-    GEExecutionEnvironment,
-    PackageInfo,
-)
-
-
-@mock.patch(
-    "great_expectations.core.usage_statistics.package_dependencies.GEExecutionEnvironment.get_all_installed_packages",
-    return_value=True,
-)
-@mock.patch("importlib.metadata.version", return_value=True)
-@mock.patch(
-    "great_expectations.core.usage_statistics.package_dependencies.GEDependencies.get_dev_dependency_names",
-    return_value=True,
-)
-@mock.patch(
-    "great_expectations.core.usage_statistics.package_dependencies.GEDependencies.get_required_dependency_names",
-    return_value=True,
-)
-def test_get_installed_packages(
-    get_required_dependency_names,
-    get_dev_dependency_names,
-    mock_version,
-    get_all_installed_packages,
-):
-    get_required_dependency_names.return_value = ["req-package-1", "req-package-2"]
-    get_dev_dependency_names.return_value = ["dev-package-1", "dev-package-2"]
-    mock_version.return_value = "8.8.8"
-    get_all_installed_packages.return_value = [
-        "req-package-1",
-        "req-package-2",
-        "dev-package-1",
-        "dev-package-2",
-    ]
-
-    ge_execution_environment: GEExecutionEnvironment = GEExecutionEnvironment()
-
-    assert ge_execution_environment.installed_required_dependencies == [
-        PackageInfo(
-            package_name="req-package-1",
-            installed=True,
-            version=version.Version("8.8.8"),
-        ),
-        PackageInfo(
-            package_name="req-package-2",
-            installed=True,
-            version=version.Version("8.8.8"),
-        ),
-    ]
-
-    assert ge_execution_environment.installed_dev_dependencies == [
-        PackageInfo(
-            package_name="dev-package-1",
-            installed=True,
-            version=version.Version("8.8.8"),
-        ),
-        PackageInfo(
-            package_name="dev-package-2",
-            installed=True,
-            version=version.Version("8.8.8"),
-        ),
-    ]
-
-
-@mock.patch(
-    "great_expectations.core.usage_statistics.package_dependencies.GEExecutionEnvironment.get_all_installed_packages",
-    return_value=True,
-)
-@mock.patch(
-    "importlib.metadata.version",
-    return_value=True,
-)
-@mock.patch(
-    "great_expectations.core.usage_statistics.package_dependencies.GEDependencies.get_dev_dependency_names",
-    return_value=True,
-)
-@mock.patch(
-    "great_expectations.core.usage_statistics.package_dependencies.GEDependencies.get_required_dependency_names",
-    return_value=True,
-)
-def test_get_not_installed_packages(
-    get_required_dependency_names,
-    get_dev_dependency_names,
-    mock_version,
-    get_all_installed_packages,
-):
-    """
-    This test raises metadata.PackageNotFoundError when calling importlib.metadata.version
-    which is raised when a package is not installed
-
-    """
-    get_required_dependency_names.return_value = [
-        "req-package-1",
-        "req-package-2",
-        "not-installed-req-package-1",
-        "not-installed-req-package-2",
-    ]
-    get_dev_dependency_names.return_value = [
-        "dev-package-1",
-        "dev-package-2",
-        "not-installed-dev-package-1",
-        "not-installed-dev-package-2",
-    ]
-    mock_version.return_value = "8.8.8"
-    get_all_installed_packages.return_value = [
-        "req-package-1",
-        "req-package-2",
-        "dev-package-1",
-        "dev-package-2",
-    ]
-
-    ge_execution_environment: GEExecutionEnvironment = GEExecutionEnvironment()
-
-    assert ge_execution_environment.installed_required_dependencies == [
-        PackageInfo(
-            package_name="req-package-1",
-            installed=True,
-            version=version.Version("8.8.8"),
-        ),
-        PackageInfo(
-            package_name="req-package-2",
-            installed=True,
-            version=version.Version("8.8.8"),
-        ),
-    ]
-
-    assert ge_execution_environment.installed_dev_dependencies == [
-        PackageInfo(
-            package_name="dev-package-1",
-            installed=True,
-            version=version.Version("8.8.8"),
-        ),
-        PackageInfo(
-            package_name="dev-package-2",
-            installed=True,
-            version=version.Version("8.8.8"),
-        ),
-    ]
-
-    assert ge_execution_environment.not_installed_required_dependencies == [
-        PackageInfo(
-            package_name="not-installed-req-package-1", installed=False, version=None
-        ),
-        PackageInfo(
-            package_name="not-installed-req-package-2", installed=False, version=None
-        ),
-    ]
-
-    assert ge_execution_environment.not_installed_dev_dependencies == [
-        PackageInfo(
-            package_name="not-installed-dev-package-1", installed=False, version=None
-        ),
-        PackageInfo(
-            package_name="not-installed-dev-package-2", installed=False, version=None
-        ),
-    ]
-
-
-@mock.patch(
-    "great_expectations.core.usage_statistics.package_dependencies.GEExecutionEnvironment.get_all_installed_packages",
-    return_value=True,
-)
-@mock.patch(
-    "importlib.metadata.version",
-    return_value=True,
-)
-@mock.patch(
-    "great_expectations.core.usage_statistics.package_dependencies.GEDependencies.get_dev_dependency_names",
-    return_value=True,
-)
-@mock.patch(
-    "great_expectations.core.usage_statistics.package_dependencies.GEDependencies.get_required_dependency_names",
-    return_value=True,
-)
-def test_get_installed_and_not_installed_packages(
-    get_required_dependency_names,
-    get_dev_dependency_names,
-    mock_version,
-    get_all_installed_packages,
-):
-    """
-    This test raises metadata.PackageNotFoundError when calling importlib.metadata.version
-    which is raised when a package is not installed
-
-    """
-    get_required_dependency_names.return_value = [
-        "not-installed-req-package-1",
-        "not-installed-req-package-2",
-    ]
-    get_dev_dependency_names.return_value = [
-        "not-installed-dev-package-1",
-        "not-installed-dev-package-2",
-    ]
-    mock_version.return_value = "8.8.8"
-    get_all_installed_packages.return_value = [
-        "req-package-1",
-        "req-package-2",
-        "dev-package-1",
-        "dev-package-2",
-    ]
-
-    ge_execution_environment: GEExecutionEnvironment = GEExecutionEnvironment()
-
-    assert ge_execution_environment.not_installed_required_dependencies == [
-        PackageInfo(
-            package_name="not-installed-req-package-1", installed=False, version=None
-        ),
-        PackageInfo(
-            package_name="not-installed-req-package-2", installed=False, version=None
-        ),
-    ]
-
-    assert ge_execution_environment.not_installed_dev_dependencies == [
-        PackageInfo(
-            package_name="not-installed-dev-package-1", installed=False, version=None
-        ),
-        PackageInfo(
-            package_name="not-installed-dev-package-2", installed=False, version=None
-        ),
-    ]
-
-
-def test_get_required_dependency_names_fail(ge_required_dependency_names):
-    # TODO: How to make this a valuable test?
-    ge_dependencies: GEDependencies = GEDependencies()
-    with pytest.raises(AssertionError):
-        assert (
-            ge_dependencies.get_required_dependency_names()
-            == ge_required_dependency_names.extend(["bad_dependency"])
-        )
+from great_expectations.core.usage_statistics.package_dependencies import GEDependencies
 
 
 def test__get_dependency_names():
